+++ Detail/Router/DetailRouteCase.swift
@@ -1,0 +1,10 @@
+//
+//  DetailRouteCase.swift
+//  Detail
+//
+//  Created by Muhammad Fahmi on 24/11/23.
+//
+
+public protocol MealDetailRouteCase {
+    func routeToDetail(from vc: UIViewController, withID id: String)
+}
