+++ Favorite/Router/FavoriteRouteCase.swift
@@ -1,0 +1,11 @@
+//
+//  FavoriteRouteCase.swift
+//  Favorite
+//
+//  Created by Muhammad Fahmi on 24/11/23.
+//
+
+import UIKit
+public protocol FavoriteRouteCase {
+    func routeToDetail(from vc: UIViewController, withID id: String)
+}
